new (WorkExp): JPMC Software Engineering Lite
This program simulates a day in the life of an entry level software engineer at JPMorgan Chase & Co., one of the world's oldest, largest and best-known financial institutions. With a history that traces our roots to 1799 in New York City, we carry forth the innovative spirit of our heritage firms in our global operations in over 60 countries. You’ll work with Java as you learn about git workflows, object-oriented programming, and unit testing.

This two-task program simulates the process of completing an engineering ticket for a system in the credit-card rewards department of JPMorgan Chase & Co. These tasks will walk you through the process of collecting the relevant codebase, creating a branch to make your changes, writing code, testing that code, and submitting a pull request to publish your work.